+++ packages/agent/src/actor.ts
@@ -11,8 +11,8 @@ import {
   v3ResponseBody,
 } from './agent';
 import { AgentError } from './errors';
-import { bufFromBufLike, IDL } from '@dfinity/candid';
-import { pollForResponse, PollStrategyFactory, strategy } from './polling';
+import { bufFromBufLike, IDL, uint8ToBuf } from '@dfinity/candid';
+import { pollForResponse, PollStrategyFactory, strategy, PollingOptions } from './polling';
 import { Principal } from '@dfinity/principal';
 import { RequestId } from './request_id';
 import { toHex } from './utils/buffer';
@@ -93,10 +93,9 @@ export interface CallConfig {
   agent?: Agent;
 
   /**
-   * A polling strategy factory that dictates how much and often we should poll the
-   * read_state endpoint to get the result of an update call.
+   * Options for controlling polling behavior.
    */
-  pollingStrategyFactory?: PollStrategyFactory;
+  pollingOptions?: PollingOptions;
 
   /**
    * The canister ID of this Actor.
@@ -445,7 +444,7 @@ export class Actor {
 // produces an array. Ensure that functions with single or zero return
 // values behave as expected.
 function decodeReturnValue(types: IDL.Type[], msg: ArrayBuffer) {
-  const returnValues = IDL.decode(types, Buffer.from(msg));
+  const returnValues = IDL.decode(types, msg);
   switch (returnValues.length) {
     case 0:
       return undefined;
@@ -457,7 +456,10 @@ function decodeReturnValue(types: IDL.Type[], msg: ArrayBuffer) {
 }
 
 const DEFAULT_ACTOR_CONFIG = {
-  pollingStrategyFactory: strategy.defaultStrategy,
+  pollingOptions: {
+    strategy: strategy.defaultStrategy(),
+    reuseReadStateSignatures: true,
+  },
 };
 
 export type ActorConstructor = new (config: ActorConfig) => ActorSubclass;
@@ -471,6 +473,7 @@ function _createActorMethod(
   func: IDL.FuncClass,
   blsVerify?: CreateCertificateOptions['blsVerify'],
 ): ActorMethod {
+  const encodeBuf = (str: string) => bufFromBufLike(new TextEncoder().encode(str));
   let caller: (options: CallConfig, ...args: unknown[]) => Promise<unknown>;
   if (func.annotations.includes('query') || func.annotations.includes('composite_query')) {
     caller = async (options, ...args) => {
@@ -522,7 +525,7 @@ function _createActorMethod(
       };
 
       const agent = options.agent || actor[metadataSymbol].config.agent || getDefaultAgent();
-      const { canisterId, effectiveCanisterId, pollingStrategyFactory } = {
+      const { canisterId, effectiveCanisterId, pollingOptions } = {
         ...DEFAULT_ACTOR_CONFIG,
         ...actor[metadataSymbol].config,
         ...options,
@@ -540,7 +543,7 @@ function _createActorMethod(
       let certificate: Certificate | undefined;
       if (response.body && (response.body as v3ResponseBody).certificate) {
         if (agent.rootKey == null) {
-          throw new Error('Agent is missing root key');
+          throw new AgentError('Agent root key not initialized before calling');
         }
         const cert = (response.body as v3ResponseBody).certificate;
         certificate = await Certificate.create({
@@ -549,7 +552,7 @@ function _createActorMethod(
           canisterId: Principal.from(canisterId),
           blsVerify,
         });
-        const path = [new TextEncoder().encode('request_status'), requestId];
+        const path = [encodeBuf('request_status'), requestId];
         const status = new TextDecoder().decode(
           lookupResultToBuffer(certificate.lookup([...path, 'status'])),
         );
@@ -566,12 +569,14 @@ function _createActorMethod(
             const rejectMessage = new TextDecoder().decode(
               lookupResultToBuffer(certificate.lookup([...path, 'reject_message']))!,
             );
+
             const error_code_buf = lookupResultToBuffer(
               certificate.lookup([...path, 'error_code']),
             );
             const error_code = error_code_buf
               ? new TextDecoder().decode(error_code_buf)
               : undefined;
+
             throw new UpdateCallRejectedError(
               cid,
               methodName,
@@ -583,7 +588,7 @@ function _createActorMethod(
             );
           }
         }
-      } else if (response.body && 'reject_message' in response.body) {
+      } else if (response.body && (response.body as v2ResponseBody).reject_code !== undefined) {
         // handle v2 response errors by throwing an UpdateCallRejectedError object
         const { reject_code, reject_message, error_code } = response.body as v2ResponseBody;
         throw new UpdateCallRejectedError(
@@ -599,9 +604,12 @@ function _createActorMethod(
 
       // Fall back to polling if we receive an Accepted response code
       if (response.status === 202) {
-        const pollStrategy = pollingStrategyFactory();
+        const pollOptions: PollingOptions = {
+          ...pollingOptions,
+          blsVerify,
+        };
         // Contains the certificate and the reply from the boundary node
-        const response = await pollForResponse(agent, ecid, requestId, pollStrategy, blsVerify);
+        const response = await pollForResponse(agent, ecid, requestId, pollOptions);
         certificate = response.certificate;
         reply = response.reply;
       }